fix: rename the file by its hooks name
